feat: 콘텐츠 타입 필터를 컴포넌트로 분리하고 학습 페이지 layout에서 제거
