+++ src/app/learn/reading/page.tsx
@@ -2,6 +2,7 @@
 
 import ArticlePreview from '@/components/ArticlePreview';
 import { useReadingContents } from '@/api/hooks/usePreview';
+import ContentTypeFilter from '@/components/ContentTypeFilter';
 
 export default function ReadingPage() {
   const {
@@ -12,22 +13,21 @@ export default function ReadingPage() {
   } = useReadingContents();
 
   if (isLoading) {
-    return <p className="mt-8">로딩 중...</p>;
+    return <p className="">로딩 중...</p>;
   }
 
   if (isError) {
-    return (
-      <p className="mt-8 text-red-500">에러가 발생했습니다: {error.message}</p>
-    );
+    return <p className="text-red-500">에러가 발생했습니다: {error.message}</p>;
   }
 
   if (!readingContents || readingContents.data.contents.length === 0) {
-    return <p className="mt-8">콘텐츠가 없습니다.</p>;
+    return <p className="">콘텐츠가 없습니다.</p>;
   }
 
   return (
-    <div className="mt-8">
-      <ul className="flex flex-col gap-4">
+    <div className="">
+      <ContentTypeFilter />
+      <ul className="flex flex-col gap-4 mt-8">
         {readingContents.data.contents.map((content) => (
           <ArticlePreview key={content.contentId} data={content} />
         ))}
